# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "What Deck API",
        "description": "Magic: The Gathering card collection management API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/collections": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "List all collections",
                "description": "Get a paginated list of card collections",
                "operationId": "7b241eefae469d00143adf95e94029b1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Collection"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Create a new collection",
                "description": "Create a new card collection",
                "operationId": "e22d3436e9fcb5530d8a74cb9f124037",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "name"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "My Main Collection"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Collection of rare and valuable cards",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collection created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collections/{id}": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Get a specific collection",
                "description": "Retrieve a specific collection with its relationships",
                "operationId": "5346e26517da2213584c2e3c0ea2798c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Collections"
                ],
                "summary": "Update a collection",
                "description": "Update an existing collection",
                "operationId": "d353584391ab8d68997dce8bd03fdb14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Collection Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Collection updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Collections"
                ],
                "summary": "Delete a collection",
                "description": "Delete a collection (only if no card instances exist)",
                "operationId": "32002b244bf5d7d47b4fd5033a6e7414",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collection deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cannot delete collection with card instances",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/collections/{id}/card-instances": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "Get collection card instances",
                "description": "Get all card instances in a specific collection",
                "operationId": "5257271b071bf1d15abe7f82c340f1aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/PaginatedResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/CardInstance"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Paginated Response",
                "description": "Standard paginated response format",
                "properties": {
                    "data": {
                        "description": "Array of resource items",
                        "type": "array",
                        "items": {}
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 5
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 15
                            },
                            "total": {
                                "type": "integer",
                                "example": 67
                            }
                        },
                        "type": "object"
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=5"
                            },
                            "prev": {
                                "type": "string",
                                "format": "url",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "url",
                                "example": "https://api.example.com/cards?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "title": "Validation Error Response",
                "description": "Response format for validation errors",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "title": [
                                "The title field is required."
                            ],
                            "type": [
                                "The type field is required."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "card": [
                                "This card has existing instances and cannot be deleted"
                            ]
                        },
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "Success Response",
                "description": "Standard success response format",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully"
                    }
                },
                "type": "object"
            },
            "CardInstance": {
                "title": "Card Instance Resource",
                "description": "Individual physical card instance resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "card_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "collection_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "deck_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "mint",
                            "near_mint",
                            "lightly_played",
                            "moderately_played",
                            "heavily_played",
                            "damaged"
                        ],
                        "example": "near_mint"
                    },
                    "foil": {
                        "type": "boolean",
                        "example": false
                    },
                    "language": {
                        "type": "string",
                        "example": "English"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "red",
                            "burn",
                            "instant"
                        ],
                        "nullable": true
                    },
                    "purchase_price": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5,
                        "nullable": true
                    },
                    "alter": {
                        "type": "boolean",
                        "example": false
                    },
                    "proxy": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "card": {
                        "$ref": "#/components/schemas/Card"
                    },
                    "collection": {
                        "$ref": "#/components/schemas/Collection"
                    },
                    "deck": {
                        "$ref": "#/components/schemas/Deck"
                    },
                    "is_in_deck": {
                        "description": "Whether this instance is assigned to a deck",
                        "type": "boolean",
                        "example": true
                    },
                    "is_available": {
                        "description": "Whether this instance is available (not in a deck)",
                        "type": "boolean",
                        "example": false
                    },
                    "condition_description": {
                        "description": "Human-readable condition description",
                        "type": "string",
                        "example": "Near Mint (NM)"
                    }
                },
                "type": "object"
            },
...
            "Collection": {
                "title": "Collection Resource",
                "description": "Card collection resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "My Main Collection"
                    },
                    "description": {
                        "type": "string",
                        "example": "Collection of rare and valuable cards",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "card_instances": {
                        "description": "Card instances in this collection (only when relationship is loaded)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CardInstance"
                        }
                    },
                    "total_card_instances": {
                        "description": "Total number of card instances in collection",
                        "type": "integer",
                        "example": 150
                    },
                    "cards_in_decks": {
                        "description": "Number of cards currently assigned to decks",
                        "type": "integer",
                        "example": 60
                    },
                    "available_cards": {
                        "description": "Number of cards not assigned to decks",
                        "type": "integer",
                        "example": 90
                    },
                    "unique_cards": {
                        "description": "Number of unique card types in collection",
                        "type": "integer",
                        "example": 75
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}